requirements.txt
selenium==4.7.2
webdriver-manager==3.8.5
flask==3.0.0
pyperclip==1.8.2
python-dateutil==2.8.2